Version 1.7.30
FossilOrigin-Name: c9c3a35e677b9dc4f5e327fdc74d8a62e32a66b6d019f5e3076b0d10770e470d

+++ lib/row.h
@@ -242,7 +242,7 @@ private:
 
 public:
   Row() {}
-  Row(MYSQL_ROW d, const ResUse *r, unsigned int *jj, bool te = false) 
+  Row(MYSQL_ROW d, const ResUse *r, unsigned long *jj, bool te = false) 
     : res(r), throw_exceptions(te), initialized(false)
 		{
 		  if (!d || !r) {
